requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.12.0
chromadb==0.4.22
python-docx==1.1.0
langchain==0.1.9
python-dotenv==1.0.0
pydantic==2.6.1
pypdf2==3.0.1 